requirements.txt
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
automaton==3.2.0
autopage==0.5.2
bcrypt==4.2.0
boto3==1.35.34
botocore==1.35.34
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cliff==4.7.0
cmd2==2.4.3
cryptography==43.0.1
debtcollector==3.0.0
decorator==5.1.1
dnspython==2.7.0
dogpile.cache==1.3.3
ecdsa==0.19.0
eventlet==0.37.0
fastapi==0.115.0
fasteners==0.19
futurist==3.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iso8601==2.1.0
Jinja2==3.1.4
jmespath==1.0.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.8.0
keystonemiddleware==10.7.1
kombu==5.4.2
logutils==0.3.5
Mako==1.3.5
mangum==0.19.0
MarkupSafe==3.0.0
msgpack==1.1.0
netaddr==1.3.0
netifaces==0.11.0
networkx==3.3
openstacksdk==4.0.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.1.0
oslo.cache==3.8.0
oslo.concurrency==6.1.0
oslo.config==9.6.0
oslo.context==5.6.0
oslo.db==16.0.0
oslo.i18n==6.4.0
oslo.log==6.1.2
oslo.messaging==14.9.0
oslo.metrics==0.9.0
oslo.middleware==6.2.0
oslo.policy==4.4.0
oslo.reports==3.4.0
oslo.serialization==5.5.0
oslo.service==3.5.0
oslo.upgradecheck==2.4.0
oslo.utils==7.3.0
oslo.versionedobjects==3.4.0
packaging==24.1
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.0
pecan==1.5.1
pillow==10.4.0
platformdirs==4.3.6
prettytable==3.11.0
prometheus_client==0.21.0
psutil==6.0.0
pyasn1==0.6.1
pycadf==3.1.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyperclip==1.9.0
python-barbicanclient==7.0.0
python-cinderclient==9.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-glanceclient==4.7.0
python-heatclient==4.0.0
python-jose==3.3.0
python-keystoneclient==5.5.0
python-multipart==0.0.12
python-neutronclient==11.3.1
python-novaclient==18.7.0
python-octaviaclient==3.8.0
python-openstackclient==7.1.3
python-swiftclient==4.6.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
repoze.lru==0.7
requests==2.32.3
requestsexceptions==1.4.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
statsd==4.0.1
stevedore==5.3.0
taskflow==5.9.1
tenacity==9.0.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
vine==5.1.0
warlock==2.0.1
wcwidth==0.2.13
WebOb==1.8.8
Werkzeug==3.0.4
wrapt==1.16.0
WSME==0.12.1
yappi==1.6.0
zipp==3.20.2
